// Frontend/jump-park-rh/src/types/apps/userTypes.ts
// ** Types
import { ThemeColor } from 'src/@core/layouts/types'

export type UsersType = {
  id: number
  role: string
  email: string
  status: string
  avatar: string
  company: string
  country: string
  contact: string
  full_name: string
  is_superuser: Boolean
  username: string
  currentPlan: string
  avatarColor?: ThemeColor
  picture: string
}

export type ManagersType = {
  id: number
  role: string
  email: string
  is_superuser: boolean
  full_name: string
  status: string
  avatar: string
  company: string
  country: string
  contact: string
  fullName: string
  username: string
  currentPlan: string
  avatarColor?: ThemeColor
  picture: string
}

export type ProjectListDataType = {
  id: number
  img: string
  hours: string
  totalTask: string
  projectType: string
  projectTitle: string
  progressValue: number
  progressColor: ThemeColor
}